refactor phase II
